refactor: extract actions into a specific component
