+++ src/components/creditNote/CreditNoteFormCalculation.tsx
@@ -1,16 +1,14 @@
 import { gql } from '@apollo/client'
-import { InputAdornment } from '@mui/material'
 import { FormikProps } from 'formik'
 import { debounce } from 'lodash'
 import _get from 'lodash/get'
 import { useCallback, useEffect, useMemo } from 'react'
-import styled, { css } from 'styled-components'
 import { array, number, object, string } from 'yup'
 
+import { CreditNoteActions } from '~/components/creditNote/CreditNoteActions'
 import { CreditNoteEstimationLine } from '~/components/creditNote/CreditNoteEstimationLine'
-import { Alert, Button, Skeleton, Tooltip, Typography } from '~/components/designSystem'
-import { AmountInputField, ComboBox, ComboBoxField } from '~/components/form'
-import { getCurrencySymbol, intlFormatNumber } from '~/core/formats/intlFormatNumber'
+import { Alert } from '~/components/designSystem'
+import { intlFormatNumber } from '~/core/formats/intlFormatNumber'
 import { deserializeAmount, getCurrencyPrecision } from '~/core/serializers/serializeAmount'
 import {
   CreditNoteItemInput,
@@ -22,7 +20,6 @@ import {
 } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
 import { DEBOUNCE_SEARCH_MS } from '~/hooks/useDebouncedSearch'
-import { theme } from '~/styles'
 
 import { CreditNoteForm, CreditTypeEnum, PayBackErrorEnum } from './types'
 
@@ -91,6 +88,8 @@ export const CreditNoteFormCalculation = ({
   const canOnlyCredit =
     invoice?.paymentStatus !== InvoicePaymentStatusTypeEnum.Succeeded ||
     !!invoice.paymentDisputeLostAt
+  const canRefund = !canOnlyCredit
+
   const currency = invoice?.currency || CurrencyEnum.Usd
   const currencyPrecision = getCurrencyPrecision(currency)
   const isLegacyInvoice = (invoice?.versionNumber || 0) < 3
@@ -229,7 +228,7 @@ export const CreditNoteFormCalculation = ({
 
   return (
     <div>
-      <CalculationContainer>
+      <div className="ml-auto flex w-full max-w-100 flex-col gap-3">
         {hasCouponLine && (
           <CreditNoteEstimationLine
             label={translate('text_644b9f17623605a945cafdbb')}
@@ -327,212 +326,19 @@ export const CreditNoteFormCalculation = ({
             }
           />
         )}
-      </CalculationContainer>
-      {!canOnlyCredit && (
-        <PayBackBlock>
-          <PayBackLine $multiline={payBack.length > 1}>
-            <ComboBox
-              name="payBack.0.type"
-              value={payBack[0]?.type}
-              onChange={(value) => {
-                if (value === CreditTypeEnum.refund && hasCreditOrCoupon) {
-                  formikProps.setFieldValue('payBack', [
-                    {
-                      type: value,
-                      value: Number(invoice?.refundableAmountCents || 0) / 100,
-                    },
-                    {
-                      type: CreditTypeEnum.credit,
-                      value:
-                        Math.round(
-                          (totalTaxIncluded || 0) * 100 -
-                            Number(invoice?.refundableAmountCents || 0),
-                        ) / 100,
-                    },
-                  ])
-                } else {
-                  formikProps.setFieldValue('payBack.0.type', value)
-                }
-              }}
-              placeholder={translate('text_637d0e628762bd8fc95f045d')}
-              data={[
-                {
-                  value: CreditTypeEnum?.credit,
-                  label: translate('text_637d0e720ace4ea09aaf0630'),
-                  disabled: payBack[1]?.type === CreditTypeEnum.credit,
-                },
-                {
-                  value: CreditTypeEnum?.refund,
-                  disabled: payBack[1]?.type === CreditTypeEnum.refund,
-                  label: translate(
-                    hasCreditOrCoupon
-                      ? 'text_637d10c83077eff6e8c79cd0'
-                      : 'text_637d0e6d94c87b04785fc6d2',
-                    {
-                      max: intlFormatNumber(
-                        deserializeAmount(invoice?.refundableAmountCents || 0, currency),
-                        {
-                          currency,
-                        },
-                      ),
-                    },
-                  ),
-                },
-              ]}
-            />
-            {payBack.length > 1 ? (
-              <>
-                <Tooltip
-                  title={translate('text_637e23e47a15bf0bd71e0d03', {
-                    max: intlFormatNumber(deserializeAmount(maxRefundableAmountCents, currency), {
-                      currency,
-                    }),
-                  })}
-                  placement="top-end"
-                  disableHoverListener={
-                    _get(formikProps.errors, 'payBack.0.value') !== PayBackErrorEnum.maxRefund
-                  }
-                >
-                  <AmountInputField
-                    className="max-w-38 [&_input]:text-right"
-                    name="payBack.0.value"
-                    currency={currency}
-                    formikProps={formikProps}
-                    beforeChangeFormatter={['positiveNumber']}
-                    displayErrorText={false}
-                    InputProps={{
-                      startAdornment: (
-                        <InputAdornment position="start">
-                          {getCurrencySymbol(currency)}
-                        </InputAdornment>
-                      ),
-                    }}
-                  />
-                </Tooltip>
-                <Tooltip title={translate('text_637d2e7e5af40c52246b1a12')} placement="top-end">
-                  <Button
-                    icon="trash"
-                    variant="quaternary"
-                    size="small"
-                    onClick={() =>
-                      formikProps.setFieldValue('payBack', [
-                        { type: payBack[1].type, value: totalTaxIncluded },
-                      ])
-                    }
-                  />
-                </Tooltip>
-              </>
-            ) : (
-              <Typography color="grey700">
-                {estimationLoading ? (
-                  <Skeleton variant="text" className="w-22" />
-                ) : !payBack[0]?.value || hasError ? (
-                  '-'
-                ) : (
-                  intlFormatNumber(payBack[0]?.value || 0, {
-                    currency,
-                  })
-                )}
-              </Typography>
-            )}
-          </PayBackLine>
+      </div>
 
-          {payBack.length < 2 ? (
-            <Button
-              variant="quaternary"
-              startIcon="plus"
-              onClick={() => {
-                formikProps.setFieldValue('payBack.1', {
-                  type: payBack[0]?.type
-                    ? payBack[0]?.type === CreditTypeEnum.credit
-                      ? CreditTypeEnum.refund
-                      : CreditTypeEnum.credit
-                    : undefined,
-                  value:
-                    payBack[0]?.value && (totalTaxIncluded || 0) - payBack[0]?.value
-                      ? (totalTaxIncluded || 0) - payBack[0]?.value
-                      : undefined,
-                })
-              }}
-            >
-              {translate('text_637d0e9729bcc6bb0cb77141')}
-            </Button>
-          ) : (
-            <PayBackLine $multiline>
-              <ComboBoxField
-                name="payBack.1.type"
-                formikProps={formikProps}
-                placeholder={translate('text_637d0e628762bd8fc95f045d')}
-                data={[
-                  {
-                    value: CreditTypeEnum?.credit,
-                    label: translate('text_637d0e720ace4ea09aaf0630'),
-                    disabled: payBack[0]?.type === CreditTypeEnum.credit,
-                  },
-                  {
-                    value: CreditTypeEnum?.refund,
-                    disabled: payBack[0]?.type === CreditTypeEnum.refund,
-                    label: translate(
-                      hasCreditOrCoupon
-                        ? 'text_637d10c83077eff6e8c79cd0'
-                        : 'text_637d0e6d94c87b04785fc6d2',
-                      {
-                        max: intlFormatNumber(
-                          deserializeAmount(invoice?.refundableAmountCents || 0, currency),
-                          {
-                            currency,
-                          },
-                        ),
-                      },
-                    ),
-                  },
-                ]}
-              />
-              <Tooltip
-                title={translate('text_637e23e47a15bf0bd71e0d03', {
-                  max: intlFormatNumber(
-                    deserializeAmount(invoice?.refundableAmountCents || 0, currency),
-                    {
-                      currency,
-                    },
-                  ),
-                })}
-                placement="top-end"
-                disableHoverListener={
-                  _get(formikProps.errors, 'payBack.1.value') !== PayBackErrorEnum.maxRefund
-                }
-              >
-                <AmountInputField
-                  className="max-w-38 [&_input]:text-right"
-                  name="payBack.1.value"
-                  currency={currency}
-                  formikProps={formikProps}
-                  beforeChangeFormatter={['positiveNumber']}
-                  displayErrorText={false}
-                  InputProps={{
-                    startAdornment: (
-                      <InputAdornment position="start">
-                        {getCurrencySymbol(currency)}
-                      </InputAdornment>
-                    ),
-                  }}
-                />
-              </Tooltip>
-              <Tooltip title={translate('text_637d2e7e5af40c52246b1a12')} placement="top-end">
-                <Button
-                  icon="trash"
-                  variant="quaternary"
-                  size="small"
-                  onClick={() => {
-                    formikProps.setFieldValue('payBack', [
-                      { type: payBack[0].type, value: totalTaxIncluded },
-                    ])
-                  }}
-                />
-              </Tooltip>
-            </PayBackLine>
-          )}
-        </PayBackBlock>
+      {canRefund && (
+        <CreditNoteActions
+          invoice={invoice}
+          formikProps={formikProps}
+          hasCreditOrCoupon={hasCreditOrCoupon}
+          maxRefundableAmountCents={maxRefundableAmountCents}
+          totalTaxIncluded={totalTaxIncluded}
+          currency={currency}
+          estimationLoading={estimationLoading}
+          hasError={hasError}
+        />
       )}
 
       {_get(formikProps.errors, 'payBack.0.value') === LagoApiError.DoesNotMatchItemAmounts && (
@@ -547,42 +353,3 @@ export const CreditNoteFormCalculation = ({
     </div>
   )
 }
-
-const CalculationContainer = styled.div`
-  max-width: 400px;
-  margin-left: auto;
-
-  > *:not(:last-child) {
-    margin-bottom: ${theme.spacing(3)};
-  }
-`
-
-const PayBackLine = styled.div<{ $multiline?: boolean }>`
-  display: flex;
-  align-items: center;
-
-  > *:not(:last-child) {
-    margin-right: ${theme.spacing(3)};
-  }
-
-  > *:first-child {
-    flex: 1;
-    max-width: 456px;
-  }
-
-  ${({ $multiline }) =>
-    !$multiline &&
-    css`
-      > *:last-child {
-        margin-left: auto;
-      }
-    `}
-`
-
-const PayBackBlock = styled.div`
-  margin-top: ${theme.spacing(6)};
-
-  > *:first-child {
-    margin-bottom: ${theme.spacing(6)};
-  }
-`
